:building_construction: paginas independientes registro/login
ahora el registro de usuiarios y el login son paginas independientes que ahora trabajan con react-router
